backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart
pytesseract
opencv-python-headless
